fix: `StackOverflowException` when formatting a recursive object
