--- Tests/aweXpect.Core.Tests/Formatting/Formatters/CollectionFormatterTests.cs
@@ -1,38 +1,0 @@
-﻿using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-
-// ReSharper disable PossibleMultipleEnumeration
-
-namespace aweXpect.Core.Tests.Formatting.Formatters;
-
-public sealed class CollectionFormatterTests
-{
-	[Fact]
-	public async Task ShouldFormatItems()
-	{
-		string expectedResult = "[\"1\", \"2\", \"3\", \"4\"]";
-		IEnumerable<string> value = Enumerable.Range(1, 4).Select(x => x.ToString());
-		StringBuilder sb = new();
-
-		string result = Formatter.Format(value);
-		Formatter.Format(sb, value);
-
-		await That(result).Should().Be(expectedResult);
-		await That(sb.ToString()).Should().Be(expectedResult);
-	}
-
-	[Fact]
-	public async Task ShouldLimitTo10Items()
-	{
-		string expectedResult = "[1, 2, 3, 4, 5, 6, 7, 8, 9, 10, …]";
-		IEnumerable<int> value = Enumerable.Range(1, 20);
-		StringBuilder sb = new();
-
-		string result = Formatter.Format(value);
-		Formatter.Format(sb, value);
-
-		await That(result).Should().Be(expectedResult);
-		await That(sb.ToString()).Should().Be(expectedResult);
-	}
-}
