+++ Tests/aweXpect.Core.Tests/Formatting/ValueFormatters.EnumTests.cs
@@ -1,0 +1,44 @@
+﻿using System.Text;
+
+namespace aweXpect.Core.Tests.Formatting;
+
+public partial class ValueFormatters
+{
+	public sealed class EnumTests
+	{
+		public enum Dummy
+		{
+			Foo,
+			Bar
+		}
+
+		[Theory]
+		[InlineData(Dummy.Foo, "Foo")]
+		[InlineData(Dummy.Bar, "Bar")]
+		[InlineData(null, "<null>")]
+		public async Task NullableShouldUseStringRepresentation(Dummy? value, string expectedResult)
+		{
+			StringBuilder sb = new();
+
+			string result = Formatter.Format(value);
+			Formatter.Format(sb, value);
+
+			await That(result).Should().Be(expectedResult);
+			await That(sb.ToString()).Should().Be(expectedResult);
+		}
+
+		[Theory]
+		[InlineData(Dummy.Foo, "Foo")]
+		[InlineData(Dummy.Bar, "Bar")]
+		public async Task ShouldUseStringRepresentation(Dummy value, string expectedResult)
+		{
+			StringBuilder sb = new();
+
+			string result = Formatter.Format(value);
+			Formatter.Format(sb, value);
+
+			await That(result).Should().Be(expectedResult);
+			await That(sb.ToString()).Should().Be(expectedResult);
+		}
+	}
+}
