+++ Tests/aweXpect.Core.Tests/Formatting/ValueFormatters.StringTests.cs
@@ -1,0 +1,23 @@
+﻿using System.Text;
+
+namespace aweXpect.Core.Tests.Formatting;
+
+public partial class ValueFormatters
+{
+	public sealed class StringTests
+	{
+		[Fact]
+		public async Task Strings_ShouldUseDoubleQuotationMarks()
+		{
+			string value = "foo";
+			string expectedResult = "\"foo\"";
+			StringBuilder sb = new();
+
+			string result = Formatter.Format(value);
+			Formatter.Format(sb, value);
+
+			await That(result).Should().Be(expectedResult);
+			await That(sb.ToString()).Should().Be(expectedResult);
+		}
+	}
+}
